+++ packages/koni_jxl_flutter/lib/koni_jxl_flutter.dart
@@ -2,7 +2,7 @@
 library;
 
 export 'package:koni_jxl/koni_jxl.dart';
-
 export 'src/jxl_animation_view.dart';
 export 'src/jxl_codec.dart';
 export 'src/jxl_image_provider.dart';
+export 'src/jxl_progressive_image.dart';